changed to long long

+++ hashTchaining.c
@@ -7,36 +7,36 @@
 HASHTABLE firstHashTable;
 HASHTABLE* hashtable;
 
-int hash(int x) {
-    int a = 11, b = 13;
+long long hash(long long x) {
+    long long a = 11, b = 13;
     return (a*x + b) % hashtable->size;
 }
 
-int isPrime(int x) {
-    for(int i = 2; i < sqrt(x); i++)
+long long isPrime(long long x) {
+    for(long long i = 2; i < sqrt(x); i++)
         if(x % i == 0)
             return 0;
     return 1;
 }
 
-int findNextPrime(int x) {
+long long findNextPrime(long long x) {
     while(!isPrime(x++));
     return --x;
 }
 
 void printHashTable() {
-    for (int i = 0; i < hashtable->size; ++i) {
+    for (long long i = 0; i < hashtable->size; ++i) {
         NODE* tmp = hashtable->table[i];
-        printf("%d: ", i);
+        printf("%lli: ", i);
         while (tmp != NULL) {
-            printf("%d -> ", tmp->data);
+            printf("%lli -> ", tmp->data);
             tmp = tmp->next;
         }
         printf("\n");
     }
 }
 
-void addToLinkedList(NODE** table, int hashed, int data) {
+void addToLinkedList(NODE** table, long long hashed, long long data) {
     //if no collision
     if(table[hashed]->data == 0)
         table[hashed]->data = data;
@@ -60,30 +60,30 @@ void addToLinkedList(NODE** table, int hashed, int data) {
     hashtable->alpha = ((hashtable->nOfElements) * 1.0) / hashtable->size;
 }
 
-void createNewHashTable(int number) {
-    int oldSize = hashtable->size;
+void createNewHashTable(long long number) {
+    long long oldSize = hashtable->size;
     //reset global table fields
     hashtable->nOfElements = hashtable->nOfCollisions = 0;
     hashtable->alpha = 0.0;
     //count size of new table
-    int newSize = (hashtable->size)*2;
+    long long newSize = (hashtable->size)*2;
     newSize = findNextPrime(newSize);
     hashtable->size = newSize;
 
     //create new table
     NODE** newHashTable = (NODE**)calloc(newSize, sizeof(NODE*));
-    for (int i = 0; i < newSize; ++i)
+    for (long long i = 0; i < newSize; ++i)
         newHashTable[i] = (NODE*)calloc(1,sizeof(NODE));
-    //printf("\nCreated new table with size %d, old one had %d\n\n", newSize, oldSize);
+    //prlong longf("\nCreated new table with size %d, old one had %d\n\n", newSize, oldSize);
 
     //also we have to add recent number
     addToLinkedList(newHashTable, hash(number), number);
 
     //now rehash old table and put all nums to new table
-    for (int j = 0; j < oldSize; ++j) {
+    for (long long j = 0; j < oldSize; ++j) {
         NODE* tmp = hashtable->table[j];
         while (tmp != NULL && tmp->data != 0) {
-            int hashed = hash(tmp->data);
+            long long hashed = hash(tmp->data);
             addToLinkedList(newHashTable, hashed, tmp->data);
             tmp = tmp->next;
         }
@@ -92,8 +92,8 @@ void createNewHashTable(int number) {
     hashtable->table = newHashTable;
 }
 
-void hashT_insert(int number) {
-    int hashed = hash(number);
+void hashT_insert(long long number) {
+    long long hashed = hash(number);
 
     //if load factor is quite normal
     if (hashtable->alpha < 0.5) {
@@ -101,24 +101,24 @@ void hashT_insert(int number) {
     }
         //factor is large, we need to create new hashtable
     else {
-        //printHashTable();
+        //prlong longHashTable();
         //we pass number, to add recent also
         createNewHashTable(number);
     }
 }
 
-int hashT_search(int number) {
-    int hashed = hash(number);
+long long hashT_search(long long number) {
+    long long hashed = hash(number);
 
     NODE* tmp = hashtable->table[hashed];
     while (tmp->data != number && tmp->next != NULL)
         tmp = tmp->next;
 
     if(tmp->data == number) {
-        //printf("Found number %d\n", number);
+        //prlong longf("Found number %d\n", number);
         return 1;
     } else {
-        //printf("Number %d was not found\n", number);
+        //prlong longf("Number %d was not found\n", number);
         return 0;
     }
 }
@@ -130,7 +130,7 @@ void hashT_initialize() {
     hashtable->size = 23;
     hashtable->table = (NODE**)calloc(hashtable->size, sizeof(NODE*));
     hashtable->alpha = 0;
-    for (int i = 0; i < hashtable->size; ++i)
+    for (long long i = 0; i < hashtable->size; ++i)
         hashtable->table[i] = (NODE*)calloc(1,sizeof(NODE));
 }
